Rethinking models entirely.
I had been modeling NOW damage as if it were yield (what I was used to
in my last analysis). Yield is a complex quantity, so we needed a bunch
of effects to account for the different factors that affect it. NOW
dmg, however, is caused solely by NOW! By including year and block
(etc) in the model, I was implicitly using NOW counts. Thus my variable
for NOW trap catch was next to negligible!